Move configuration.rb to configuration folder
