+++ lib/broadside.rb
@@ -4,7 +4,7 @@ require 'aws-sdk'
 
 require 'broadside/error'
 require 'broadside/logging_utils'
-require 'broadside/configuration'
+require 'broadside/configuration/configuration'
 require 'broadside/configuration/aws_config'
 require 'broadside/configuration/ecs_config'
 require 'broadside/command'
